crit points: additional constraint for higher CPs
We add an additional constraint to make sure we find higher critical
points. The constraint asks for the vector fields X_i to be non-
vanishing.
In addition the solution now only asks for the component vector. This
allows the solution to be moved from the GPU back to the host and
stored on disk.# Please enter the commit message for your changes. Lines
starting